Tidying up

+++ 03/main_03.cs
@@ -2,5 +2,5 @@
 	public Day03(int day_ref_int) : base(day_ref_int) { }
 	public Day03() : base(3) { }
 
-	private static int GetPriority(char c) => (int)c - (char.IsUpper(c) ? 38 : 96);
+	private static int GetPriority(char c) => c - (char.IsUpper(c) ? 38 : 96);
 }
